Validation for device code removed, added prefix control over generate license

+++ app/Http/Controllers/ApiLicenseController.php
@@ -6,9 +6,21 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Request;
 use Log;
+use Auth;
 
 class ApiLicenseController extends Controller {
 
+
+	/**
+	 * Create a new controller instance.
+	 *
+	 * @return void
+	 */
+	public function __construct()
+	{
+		$this->middleware('auth');
+	}
+	
 	/**
 	 * Display a listing of the resource.
 	 *
@@ -16,8 +28,14 @@ class ApiLicenseController extends Controller {
 	 */
 	public function index()
 	{
-		$licenses = ActCode::all();
-		return response()->json($licenses);
+		if (Auth::check()) {
+			$licenses = ActCode::all();
+			return response()->json($licenses);
+		} else {
+			return response()->json(array(
+				'message'	=> "please login first"
+				));
+		}
 	}
 
 	/**
@@ -61,20 +79,27 @@ class ApiLicenseController extends Controller {
         // validate
         // read more on validation at http://laravel.com/docs/validation
         // $licenses = Request::get('licensedata');
-        $validator = Validator::make(Request::all(), ApiLicenseController::rules());
-		// Log::error($licenses[0]);
 
-        if ($validator->fails()) {
+		if (Auth::check()) {
+	        $validator = Validator::make(Request::all(), ApiLicenseController::rules());
+			// Log::error($licenses[0]);
+
+	        if ($validator->fails()) {
+				return response()->json(array(
+					'success'	=> false,
+					'status'	=> $validator->messages()
+					));
+	        } else {
+				$license = ActCode::create(Request::all());
+				return response()->json(array(
+					'success'	=> true,
+					'license'	=> $license,
+					'status'	=> $validator->messages()
+					));
+			}
+		} else {
 			return response()->json(array(
-				'success'	=> false,
-				'status'	=> $validator->messages()
-				));
-        } else {
-			$license = ActCode::create(Request::all());
-			return response()->json(array(
-				'success'	=> true,
-				'license'	=> $license,
-				'status'	=> $validator->messages()
+				'message'	=> "please login first"
 				));
 		}
 	}
@@ -109,30 +134,39 @@ class ApiLicenseController extends Controller {
 	 */
 	public function update($id)
 	{	
-        // validate
-        // read more on validation at http://laravel.com/docs/validation
-        $validator = Validator::make(Request::all(), ApiLicenseController::rules($id));
+		if (Auth::check()) {
+	        // validate
+	        // read more on validation at http://laravel.com/docs/validation
+	        $validator = Validator::make(Request::all(), ApiLicenseController::rules($id
+	        	// , ['device_code'	=> 'unique:tsqgeointel_activation,device_code' . ($id ? ",$id" : '')]
+	        	// adds validation when edit mode
+	        	));
 
-        if ($validator->fails()) {
+	        if ($validator->fails()) {
+				return response()->json(array(
+					'success'	=> false,
+					'status'	=> $validator->messages()
+					));
+	        } else {
+
+				$license = ActCode::find($id);
+	            $license->act_code		= Request::get('act_code');
+	            $license->organization	= Request::get('organization');
+	            $license->status		= Request::get('status');
+	            $license->device_code	= Request::get('device_code');
+	            $license->project		= Request::get('project');
+	            $license->act_date		= Request::get('act_date');
+				$license->save();
+
+				return response()->json(array(
+					'success'	=> true,
+					'license'	=> $license,
+					'status'	=> $validator->messages()
+					));
+			}
+		} else {
 			return response()->json(array(
-				'success'	=> false,
-				'status'	=> $validator->messages()
-				));
-        } else {
-
-			$license = ActCode::find($id);
-            $license->act_code		= Request::get('act_code');
-            $license->organization	= Request::get('organization');
-            $license->status		= Request::get('status');
-            $license->device_code	= Request::get('device_code');
-            $license->project		= Request::get('project');
-            $license->act_date		= Request::get('act_date');
-			$license->save();
-
-			return response()->json(array(
-				'success'	=> true,
-				'license'	=> $license,
-				'status'	=> $validator->messages()
+				'message'	=> "please login first"
 				));
 		}
 	}
@@ -145,8 +179,14 @@ class ApiLicenseController extends Controller {
 	 */
 	public function destroy($id)
 	{
-		ActCode::destroy($id);
-		return response()->json(array('success' => true));
+		if (Auth::check()) {
+			ActCode::destroy($id);
+			return response()->json(array('success' => true));
+		} else {
+			return response()->json(array(
+				'message'	=> "please login first"
+				));
+		}
 	}
 
 }
